core/utils.py
import yfinance as yf
import numpy as np
import matplotlib.pyplot as plt
from sklearn.preprocessing import MinMaxScaler
import os
import uuid
import random

def predict_stock(ticker):
    # ✅ Download 10 years of stock data
    df = yf.download(ticker, period="10y")

    # ✅ Handle case where no data is returned
    if df.empty:
        raise ValueError("No stock data found for this ticker. Please try a different one.")

    # ✅ Use Close prices only
    close_prices = df['Close'].values.reshape(-1, 1)

    # ✅ Normalize the data
    scaler = MinMaxScaler()
    scaled = scaler.fit_transform(close_prices)

    # ✅ Predict the next price (simulated random logic)
    last_price = close_prices[-1][0]
    predicted_price = round(float(last_price * random.uniform(0.98, 1.05)), 2)

    # ✅ Create fake predictions for plotting
    actual = close_prices[-30:].flatten()
    predicted = actual * random.uniform(0.98, 1.05)

    # ✅ Create charts directory if not exists
    chart_dir = os.path.join(os.getcwd(), 'charts')
    os.makedirs(chart_dir, exist_ok=True)

    # ✅ Generate unique names for the chart images
    chart_1_name = f"{uuid.uuid4().hex}_history.png"
    chart_2_name = f"{uuid.uuid4().hex}_prediction.png"
    chart_1_path = os.path.join(chart_dir, chart_1_name)
    chart_2_path = os.path.join(chart_dir, chart_2_name)

    # ✅ Plot 10-year historical chart
    plt.figure(figsize=(10, 5))
    plt.plot(close_prices, label='Closing Price')
    plt.title(f"{ticker} - 10 Year Closing Price")
    plt.xlabel("Days")
    plt.ylabel("Price")
    plt.legend()
    plt.tight_layout()
    plt.savefig(chart_1_path)
    plt.close()

    # ✅ Plot predicted vs actual (simulated)
    plt.figure(figsize=(10, 5))
    plt.plot(actual, label="Actual")
    plt.plot(predicted, label="Predicted", linestyle='--')
    plt.title(f"{ticker} - Actual vs Predicted")
    plt.xlabel("Days")
    plt.ylabel("Price")
    plt.legend()
    plt.tight_layout()
    plt.savefig(chart_2_path)
    plt.close()

    # ✅ Simulated metrics
    metrics = {
        "mse": round(random.uniform(0.1, 1.0), 3),
        "rmse": round(random.uniform(0.3, 2.0), 3),
        "r2": round(random.uniform(0.8, 1.0), 3)
    }

    return {
        "predicted_price": predicted_price,
        "metrics": metrics,
        "chart_1": f"charts/{chart_1_name}",
        "chart_2": f"charts/{chart_2_name}"
    }
